Fixed server freezing issues
So, apparently the old resource would freeze the server when players left and when writing. A bunch of shit would break it and freeze the main thread. This fixes those issues and should make the resource usable...

The client list has now been changed to a ConcurrentDictionary because I thought my locks and stuff was breaking.. Turns out wasn't the problem.

Also fixed a error where the "playerLeft" would trigger when a player isn't in the "playerData" object.