refactor: remove location data
The location strings were localised to the client thiis ment that Chinese players would be sending Chinese strrings to the interface.

I might add this back in but, it would require another file like the "reverse_weapon_hash" file for location data. This would allow the server owners to translate the locations to their prefered language.