feat: failed sse attempt
I don't think I'm smart enough to implement SSE myself.
This branch is my failed attempts. I say attempts, because I tried to implement it with KOA and in native LUA.
Neither worked.

I think the reason is because the request and response variables being passed to the SET_HTTP_HANDLER function are... Weird.

All the examples I've found online implementing SSE have access to the request's socket... FiveM doesn't.
And for some reason, the headers don't seem to send until the request is done (needs more research). I say this because while using the Node and Lua implementations, CURL handles it fine but, browsers just seem to hang on the request. If the headers were being sent when they're supposed to, this wouldn't happen. The browser should see that it's a "text/event-stream" and handle it as such. Rather than "downloading" the resource.